rawsocket, cbor, msgpack, pypy3 now supported
