+++ requirements-test.txt
@@ -1,0 +1,2 @@
+cbor
+msgpack
