+++ requirements.txt
@@ -10,6 +10,7 @@
 # `Pipfile.lock` and then regenerate `requirements*.txt`.
 ################################################################################
 
+cbor
 certifi
 six
 websocket-client
