+++ swampyer/common.py
@@ -10,8 +10,21 @@ REGISTERED_CALL_CALLBACK = 1
 SUBSCRIPTION_TOPIC = 0
 SUBSCRIPTION_CALLBACK = 1
 
+SERIALIZERS = ['json','msgpack','cbor','ubjson','flatbuffers']
+SERIALIZERS_BLACKLIST = ['ubjson','flatbuffers']
+SERIALIZER_REGISTRY = {}
+
+WEBSOCKET_SERIALIZER_JSON = 'wamp.2.json'
+WEBSOCKET_SERIALIZER_MESSAGEPACK = 'wamp.2.msgpack'
+WEBSOCKET_SERIALIZER_CBOR = 'wamp.2.cbor'
+WEBSOCKET_SERIALIZER_UBJSON = 'wamp.2.ubjson'
+WEBSOCKET_SERIALIZER_FLATBUFFERS = 'wamp.2.flatbuffers'
+
 RAWSOCKET_SERIALIZER_JSON = 1
 RAWSOCKET_SERIALIZER_MESSAGEPACK = 1
+RAWSOCKET_SERIALIZER_CBOR = 2
+RAWSOCKET_SERIALIZER_UBJSON = 3
+RAWSOCKET_SERIALIZER_FLATBUFFERS = 4
 
 RAWSOCKET_HANDSHAKE_ERRORS = {
     0: 'illegal (must not be used)',
