+++ swampyer/exceptions.py
@@ -20,5 +20,14 @@ class ExWAMPConnectionError(SwampyException):
 class ExNotImplemented(SwampyException, NotImplementedError):
     pass
 
+class ExFatalError(SwampyException):
+    pass
+
+class ExTransportParseError(ExFatalError):
+    pass
+
+class ExShutdown(SwampyException):
+    pass
+
 # Support for deprecated WAMPConnectionError class
 WAMPConnectionError = ExWAMPConnectionError
