+++ tests/common.py
@@ -1,4 +1,9 @@
-WS_PATH = 'ws://localhost:18081/ws'
+import six
+
+try:
+    import swampyer 
+except:
+    pass
 
 TICKET_USERNAME = 'user'
 TICKET_PASSWORD = 'pass'
@@ -6,6 +11,19 @@ TICKET_PASSWORD = 'pass'
 def get_password():
     with open('test_server/.crossbar/test-password.txt','r') as f:
         return f.read()
+
+def connect_service(url='ws://localhost:18080/ws',serializer_code='json'):
+    password = get_password()
+    client = swampyer.WAMPClientTicket(
+                    url=url,
+                    username=u"test",
+                    password=six.u(password),
+                    realm=u"realm1",
+                    uri_base="",
+                    serializers=[serializer_code],
+                    auto_reconnect=False,
+                ).start()
+    return client
