+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27,py36,py38
+envlist = pypy3,py27,py36,py38
 
 [testenv]
 
@@ -12,6 +12,7 @@ allowlist_externals =
 deps = 
   pytest
   -rrequirements.txt
+  -rrequirements-test.txt
 
 # change pytest tempdir and add posargs from command line
 commands =
